// src/app/core/model/event.ts
/**
 * Tzion API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1.6
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export class EventLine {
    title?: string;
    type?: string;
    startIn?: string;
    endIn?: string;
    description?: string;
    status?: string;
    createdAt?: Date;
    uploadedAt?: Date;
    deletedAt?: Date;
}